ComputerVision_1.2/requirements.txt
fastapi
uvicorn
pydantic
numpy
Pillow
opencv-python
ultralytics
torch
